feat: realtime updating the video length on playback speed change.
