+++ src/entrypoints/content/components/HoverIcon/hover.css
@@ -3,22 +3,53 @@
   top: 37%;
   left: 47%;
   opacity: 0;
-  transition: opacity 0.2s ease-in-out;
+  transition: all 0.3s ease-in-out;
   z-index: 99999;
+  width: 44px;
+  height: 44px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.neomorphic-icon-bg {
+  width: 36px;
+  height: 36px;
+  border-radius: 50%;
+  background: #e0e5ec;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  box-shadow: 3px 3px 6px rgba(163, 177, 198, 0.6),
+    -3px -3px 6px rgba(255, 255, 255, 0.8);
+  transition: all 0.3s ease;
+}
+
+#hover-icon {
+  width: 20px;
+  height: 20px;
+  color: #6b7280;
+  transition: all 0.3s ease;
 }
 
 ytd-rich-item-renderer #thumbnail:hover .hover-icon-container {
   opacity: 1;
 }
 
-.hover-icon-container #hover-icon {
-  transition: opacity 0.2s ease-in-out;
-  padding: 5px;
-  border-radius: 50%;
-  background-color: black;
+.neomorphic-icon-bg:hover {
+  box-shadow: inset 2px 2px 5px rgba(163, 177, 198, 0.6),
+    inset -2px -2px 5px rgba(255, 255, 255, 0.8);
+  transform: scale(0.95);
 }
 
-#hover-icon:hover {
-  cursor: pointer;
-  opacity: 1;
+.neomorphic-icon-bg:hover #hover-icon {
+  color: #3b82f6;
+  transform: scale(0.95);
+}
+
+/* Active/pressed state */
+.neomorphic-icon-bg:active {
+  box-shadow: inset 3px 3px 6px rgba(163, 177, 198, 0.6),
+    inset -3px -3px 6px rgba(255, 255, 255, 0.8);
+  transform: scale(0.9);
 }
